Tentative style updates to components section
