+++ src/DraggableComponentWrapper.jsx
@@ -1,0 +1,86 @@
+import { React, useRef } from 'react';
+import { useDrag, useDrop } from 'react-dnd';
+import { ItemTypes } from './ItemTypes';
+import Component from './Component';
+
+const style = {
+  cursor: 'move',
+};
+
+const DraggableComponentWrapper = ({
+  id, index, moveCard, componentType, componentValues, handleCardChange, deleteCard,
+}) => {
+  const ref = useRef(null);
+
+  const [{ handlerId }, drop] = useDrop({
+    accept: ItemTypes.CARD,
+    collect(monitor) {
+      return {
+        handlerId: monitor.getHandlerId(),
+      };
+    },
+    hover(item, monitor) {
+      if (!ref.current) {
+        return;
+      }
+      const dragIndex = item.index;
+      const hoverIndex = index;
+      // Don't replace items with themselves
+      if (dragIndex === hoverIndex) {
+        return;
+      }
+      // Determine rectangle on screen
+      const hoverBoundingRect = ref.current?.getBoundingClientRect();
+      // Get vertical middle
+      const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
+      // Determine mouse position
+      const clientOffset = monitor.getClientOffset();
+      // Get pixels to the top
+      const hoverClientY = clientOffset.y - hoverBoundingRect.top;
+      // Only perform the move when the mouse has crossed half of the items height
+      // When dragging downwards, only move when the cursor is below 50%
+      // When dragging upwards, only move when the cursor is above 50%
+      // Dragging downwards
+      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
+        return;
+      }
+      // Dragging upwards
+      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
+        return;
+      }
+      // Time to actually perform the action
+      moveCard(dragIndex, hoverIndex);
+      // Note: we're mutating the monitor item here!
+      // Generally it's better to avoid mutations,
+      // but it's good here for the sake of performance
+      // to avoid expensive index searches.
+      // eslint-disable-next-line no-param-reassign
+      item.index = hoverIndex;
+    },
+  });
+
+  const [{ isDragging }, drag] = useDrag({
+    type: ItemTypes.CARD,
+    item: () => ({ id, index }),
+    collect: (monitor) => ({
+      isDragging: monitor.isDragging(),
+    }),
+  });
+
+  const opacity = isDragging ? 0 : 1;
+
+  drag(drop(ref));
+
+  return (
+    <div ref={ref} data-handler-id={handlerId} style={{ ...style, opacity }}>
+      <Component
+        componentType={componentType}
+        componentValues={componentValues}
+        handleCardChange={handleCardChange}
+        deleteCard={deleteCard}
+      />
+    </div>
+  );
+};
+
+export default DraggableComponentWrapper;
